+++ src/tyr/matrix_serializer.cc
@@ -132,6 +132,12 @@ json::MapPtr serialize(const Api& request,
   if (options.has_id_case()) {
     json->emplace("id", options.id());
   }
+
+  // add warnings to json response
+  if (request.info().warnings_size() >= 1) {
+    json->emplace("warnings", valhalla::tyr::serializeWarnings(request));
+  }
+
   return json;
 }
 } // namespace valhalla_serializers